MINIFICPP-2217 - Refactor
